requirements.txt
Django>=1.9,<1.10
cookiecutter==1.0.0
psycopg2cffi
raven==5.0.0
markdown
django-libsass
certifi
dj-database-url
django-cas-ng==3.5.4
django-mptt
requests
elastic-git
unicore.content
wagtail==1.4.5
responses
babel
wagtailmodeladmin
wagtailmedia
django-extensions
django-celery
redis
six>=1.9
